Refactor QueryDefinition{,Tooltip} to use metabase-lib
