+++ frontend/src/metabase/reference/components/Formula.jsx
@@ -8,9 +8,8 @@ import S from "./Formula.css";
 
 import Icon from "metabase/components/Icon";
 
-import QueryDefinition from "metabase/query_builder/components/dataref/QueryDefinition";
+import QueryDefinition from "metabase/query_builder/components/QueryDefinition";
 import { fetchTableMetadata } from "metabase/redux/metadata";
-import { getMetadata } from "metabase/selectors/metadata";
 
 import type Metadata from "metabase-lib/lib/metadata/Metadata";
 
@@ -18,12 +17,8 @@ const mapDispatchToProps = {
   fetchTableMetadata,
 };
 
-const mapStateToProps = (state, props) => ({
-  metadata: getMetadata(state, props),
-});
-
 @connect(
-  mapStateToProps,
+  null,
   mapDispatchToProps,
 )
 export default class Formula extends Component {
@@ -33,7 +28,6 @@ export default class Formula extends Component {
     isExpanded: boolean,
     expandFormula: any,
     collapseFormula: any,
-    metadata: Metadata,
   };
 
   render() {
@@ -43,7 +37,6 @@ export default class Formula extends Component {
       isExpanded,
       expandFormula,
       collapseFormula,
-      metadata,
     } = this.props;
 
     return (
@@ -65,7 +58,6 @@ export default class Formula extends Component {
               <QueryDefinition
                 className={S.formulaDefinitionInner}
                 object={entity}
-                metadata={metadata}
               />
             </div>
           )}
